Mise en place du main et de la logique des objects

+++ Vector3f.cpp
@@ -3,10 +3,10 @@
 
 Vector3f::Vector3f()
 {
-	x_ = 0, y_ = 0, z_ = 0;
+	x_ = 0.0, y_ = 0.0, z_ = 0.0;
 }
 
-Vector3f::Vector3f(int x, int y, int z)
+Vector3f::Vector3f(double x, double y, double z)
 {
 	x_ = x, y_ = y, z_ = z;
 }
